+++ aio/src/app/documents/document-contents.ts
@@ -1,0 +1,4 @@
+export interface DocumentContents {
+  title: string;
+  contents: string;
+}
