+++ aio/src/app/navigation/navigation-node.ts
@@ -1,0 +1,7 @@
+export interface NavigationNode {
+  url?: string;
+  title?: string;
+  tooltip?: string;
+  target?: string;
+  children?: NavigationNode[];
+}
